refactor(tests): rename runtime_test_support → driver_test_support
The helpers are generic compile/run plumbing for the kaede driver, used
by both runtime tests and import tests, so name them after the test
target rather than after one of the consumers.
